webapp/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
geopandas==1.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
neo4j==5.28.1
networkx==3.5
numpy==2.3.1
osmnx==2.0.4
packaging==25.0
pandas==2.3.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
passlib[bcrypt]==1.7.4
